+++ cli/spiff/f2_polyrat_handlers.cpp
@@ -2,12 +2,12 @@
 
 #include "rationals/f2_polyrat_t.h"
 #include "random/f2_poly_random.h"
-#include "cliparser/cmd_line_matrix_ops.h"
-#include "cliparser/cmd_line_ops.h"
-#include "cliparser/cmd_line_vector_ops.h"
+#include "cli_parser/cmd_line_matrix_ops.h"
+#include "cli_parser/cmd_line_ops.h"
+#include "cli_parser/cmd_line_vector_ops.h"
 
 int f2_pr_op_main(int argc, char **argv, usage_t *pusage) {
-  spffl::cliparser::cmd_line_parse<spffl::rationals::f2_polyrat_t>(argc - 1,
+  spffl::cli_parser::cmd_line_parse<spffl::rationals::f2_polyrat_t>(argc - 1,
       argv + 1, spffl::rationals::f2_polyrat_t(0),
       spffl::rationals::f2_polyrat_t(1));
   return 0;
@@ -49,7 +49,7 @@ int f2_pr_random_main(int argc, char **argv, usage_t *pusage) {
 }
 
 int f2_pr_mat_op_main(int argc, char **argv, usage_t *pusage) {
-  spffl::cliparser::cmd_line_mat_parse<spffl::rationals::f2_polyrat_t>(argc - 1,
+  spffl::cli_parser::cmd_line_mat_parse<spffl::rationals::f2_polyrat_t>(argc - 1,
       argv + 1,
       spffl::rationals::f2_polyrat_t(spffl::polynomials::f2_poly_t(0)),
       spffl::rationals::f2_polyrat_t(spffl::polynomials::f2_poly_t(1)));
