+++ cli/spiff/handlers.h
@@ -1,7 +1,7 @@
 #ifndef HANDLERS_H
 #define HANDLERS_H
 
-#include "cliparser/cmd_line_ops.h"
+#include "cli_parser/cmd_line_ops.h"
 
 typedef void usage_t(char *argv0);
 typedef int main_t(int argc, char **argv, usage_t *pusage);
