+++ spffl/CMakeLists.txt
@@ -9,10 +9,10 @@ add_subdirectory(bits)
 add_subdirectory(containers)
 add_subdirectory(polynomials)
 add_subdirectory(rationals)
-add_subdirectory(q_cyclo)
+add_subdirectory(q_cyclotomic)
 add_subdirectory(random)
 add_subdirectory(factorization)
 add_subdirectory(units)
 add_subdirectory(list)
 add_subdirectory(linalg)
-add_subdirectory(cliparser)
+add_subdirectory(cli_parser)
