+++ spffl/q_cyclotomic/CMakeLists.txt
@@ -1,6 +1,6 @@
 cmake_minimum_required (VERSION 3.24)
 
-project(q_cyclo)
+project(q_cyclotomic)
 
 add_library(${PROJECT_NAME}
   qff.cpp
